+++ src/wallet.cpp
@@ -247,7 +247,7 @@ void Wallet::reload(bool refresh_accounts)
     auto handler = new GetSubAccountsHandler(m_session, refresh_accounts);
     QObject::connect(handler, &Handler::done, this, [this, handler] {
         handler->deleteLater();
-        const bool create_segwit = m_network->isElectrum();
+        const bool create_segwit = m_network->isElectrum() && (m_restoring || m_device);
         bool has_segwit = false;
 
         m_accounts.clear();
@@ -259,11 +259,14 @@ void Wallet::reload(bool refresh_accounts)
             m_accounts.append(account);
         }
 
-        if (m_restoring && create_segwit && !has_segwit) {
+        if (create_segwit && !has_segwit) {
             auto handler = new CreateAccountHandler({{ "name", "Segwit Account" }, { "type", "p2wpkh" }}, m_session);
             QObject::connect(handler, &Handler::done, this, [=] {
                 handler->deleteLater();
                 reload(false);
+            });
+            QObject::connect(handler, &Handler::resolver, this, [](Resolver* resolver) {
+                resolver->resolve();
             });
             handler->exec();
             return;
@@ -626,7 +629,7 @@ void Wallet::setSession()
     updateSettings();
     updateCurrencies();
     updateConfig();
-    reload();
+    reload(true);
     updateReady();
 }